requirements.txt
langchain-core==0.2.38
langchain-openai==0.1.23
langgraph==0.2.16

python-dotenv==1.0.1
typing-extensions==4.12.2

requests==2.31.0
httpx==0.27.0